backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.1
django-allauth==0.54.0
django-cleanup==7.0.0
django-cors-headers==4.0.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.4.0
